Consolidate the exchange-rate logic into the actions
These functions aren't really selectors, so they don't belong in the selectors file.
